backend/code/pyproject.toml
[project]
name = "grawantura"
version = "0.1.0"
description = ""
authors = [
    {name = "Dominik Dlugajczyk",email = "msocek@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.15.1,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "starlette (>=0.46.1,<0.47.0)",
    "python-decouple (>=3.8,<4.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    # "quackquack @ file://quackquack",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 99
known_first_party = [ "grawantura" ]
multi_line_output = 3
sections = [ "FUTURE", "STDLIB", "OTHER", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
force_single_line = true
default_section = "OTHER"

[tool.black]
line_length = 99
target-version = [ "py38" ]
ignore = [ "W503", "W504" ]

[tool.pytest.ini_options]
addopts = "--cov --cov-report term:skip-covered --cov-report html --no-cov-on-fail -vv --ff -x grawantura"
asyncio_mode = "auto"
markers = [
    "integration: all tests which are needing of external services, like database or appnexus api", ]
norecursedirs = [
    ".cache",
    ".git",
    ".idea",
    "bin",
    "*migrations*",
    "spa",
    "static",
    "var",
    "docker",
    "common" ]
filterwarnings = [
    "error::RuntimeWarning",
    "ignore:SelectableGroups dict interface is deprecated. Use select.:DeprecationWarning",
    "ignore:The hookimpl CovPlugin.* uses old-style configuration options.*:DeprecationWarning",
]
[tool.coverage.run]
branch = true
omit = [
    "qq/*",
    "quackquack/*",
    "*/tests/*",
    "*/test_*/*",
    "*/versions/*",
    "*/testing.py",
    "*/settings/*",
    "*/app/*",
    "*/admin.py",
    "*/.cache/*",
    "/home/user/*",
    "*/.cache/pypoetry/*",
    "conftest.py"]

[tool.poetry.group.dev.dependencies]
ipython = "^9.0.2"
ipdb = "^0.13.13"
icecream = "^2.1.4"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pytest-async = "^0.1.1"
pytest-asyncio = "^0.25.3"

